requirements.txt
gtts==2.2.4
Pillow==9.1.1
playsound==1.3.0
pynput==1.7.6
pystray==0.19.3
python_weather==0.3.7
pywin32==304
requests~=2.28.1
scikit_learn==1.1.1
SpeechRecognition==3.8.1
wikipedia==1.4.0
python-dateutil~=2.8.2
numpy~=1.24.1
scipy~=1.10.0
customtkinter~=5.0.3